[REF] sale_timesheet: make project billable wizard
When introducing the 'employee rate' mode to timesheet
on project, we break the wizard: its goal was to link the
existing timesheets to the SO line in order to invoice them.
(Employee rate commit: 50af6a561b6ff0c46838e86184c35a6da51e30a5)

This commit makes the wizard compatible with the employee rate
feature, without breaking invoicing at task rate flow.

Two possible flows, starting with a non billable project:
  * project rate: the wizard will create a SO with one SOL,
  	and will link the task without SOL set to the new one.
  	The project will also be linked to that SOL.
  * employee rate: the user has to fill a map to associate
  	product and employee. If the same product is set 2 times
  	with the same price (and so, 2 different employees), only
  	one SOL will be created on the SO. When submitting, the
  	project map will be filled and the project will be linked
  	to the SO. No fallback SOL will be set on project or on
  	the tasks.

Both use cases will be stopped if a timesheet of the project's
tasks is already linked to a sale line. If not, the timesheets
will be associated to the correct SOL (dependning of the billable
type), and the delivered quantity will be incremented, making
the SO invoicable. The tasks'customer will also be set.

Future work: setting a product per task (which will be a real
'task rate' feature).

Task #1883640