[FIX] web: use read to update record values after resequence
Commit:
https://github.com/odoo/odoo/commit/72e64b883bc5bdae398e8caa7b7debcf1db278c8
Updated the resequence logic in a way that was almost falser than before.
The added test however did work by coincidence, as index values were equal to
the sequence field values.
To be sure that we synchronize with what happens on the server, we do a read
after the resequence.
Additionnally, we take into account the result of the server resequence return;
if it is false, it means no resequencing happened, so we should not do a read.

opw 1867049

closes odoo/odoo#27184