[FIX] website: restore proper carousel options
We have JS which tooks care of customizing the color and background
options of .carousel elements so that the active carousel item is
edited instead. The recent snippet refactoring added those options on
.carousel-item elements directly, so that slide and carousel options
are separated in the customize menu. This however made the options
complex to use and some of them impossible to use. This also added color
and background options on non-snippet carousels like for the products in
the shop, which was not meant to be. This commit goes back to the old
system for now.

Note: using "[class*='s_']" as a selector to match snippets is not
correct as it matches the elements whose classes contains 's_'... and
in our case the body was matched because it has the class
'editor_ha*s_*snippets'. A proper way would be to use this selector:
"[class^='s_'], [class*=' s_']". But even then, this is not really a
good way.

task-1878245