[FIX] sale_timesheet: duplicate project remove tasks SOL
When duplicating a task, we want the copy to have
the same sale order line then the original.
However, we duplicating a project, the tasks are
copied too. But, in that case, we don't want them
to be linked to any sale line, as the copy project
will not be.

This commit fixes that use case, and provides a test.

Task #1883640