[FIX] portal, website: allow to use wide brand logo
When the user chose a very wide logo for the website navbar, it
overflew over the menu items (making the "+" feature look broken).
This in fact occurs since commit https://github.com/odoo/odoo/commit/f4735a84fce8147ee178e09514b646ef3ec7373c.
Its goal was to avoid top menu flickering due to that "+" button
feature. To do that, menu items were hidden with an height of 0 during
loading, then, once the correct number of items needed for the "+"
feature was computed, they were made visible to the user. Another
problem was due to the logo image loading. When it is not loaded yet, we
cannot properly compute the number of items which is needed. To solve
that problem, we fixed the size of the image container instead of
setting the height of the image itself.

But fixing the size of the image container only created a new bug: very
wide logos overflow this container. This commit solves all the problems:

1) The height of the image is set to be equal to the navbar links.
2) The height of the navbar is not forced, meaning that if the user
   decides to change the height of its logo to go against (1),
   everything will be adapted accordingly.
3) During loading, the menu items are still 0px tall, hidden and made
   visible once the "+" feature is enabled.
4) The "+" feature is only enabled once all the navbar images have been
   successfully loaded.

task-1878245