[FIX] web: background layout landscape no +1 page
The report layout "background" has an image background from bottom to
top. When we transform the report in PDF, we have no idea in the
document of how many page, which orientation, width or height it will
take.

Thus to be able to have the background at the bottom of a page when
there is not enough content, we set a min-height for the body content.
This height was 1000px originally and has been decreased to 900px in
7ca069e312 because an empty page was possibly added (since the height of
body + header + footer was bigger than page size).

This issue is about getting a blank page when pdf is in landscape.
For example in A4 with 90ppi has a 1052px height in portrait, but 744px
height in landscape so we always get at least two pages with possibly a
blank one.

With this commit we lower it to 620px allowing 124px headers/footers.

opw-1881059
closes #26962