[IMP] im_livechat : Fix livechat session and operator reports
Session Statistics : To keep only the operators, add a join to
im_livechat_channel_im_user that keeps the link between who is the operator
for which channel.

Operator Report : To keep only the users that have really participated
to the channel, add a join to mail_channel_partner_table that keeps the link
between who participate to which channel.

The report are only taking into account current operators.
If, on a livechat channel, an operator is removed and another user is
set as operator for ths channel. this new operator will have the stats
of the removed operator. This will stay like this until we add a field
that records who is the operator at the time the discussion. This field
should be added on mail_channel_partner table and should be a role.
So if a user is no more operator, he will still be flagged as operator of
this particular discussion.

Task ID 1883227
closes PR #26995