[FIX] web: tests: adapt JS tests from forward-port 4efe82f0ca060971ffce31cad50ec0643a7cab0c
The test was failing because it was using an old approach to mock
services in JS tests. Also, it provides ajax service in the list
of services, which was not supported in a mocked environment.

With this commit, when an ajax service is provided in the test,
it will use this service instead of using mocked RPCs from the
mocked server.