[FIX] web, website: root.widget defined even without website
Revision on https://github.com/odoo/odoo/commit/a72dd5d9a4f1b8f08016d73f873fbc74bcb86700

With the commit above, the following error was shown in the devtools
when accessing a frontend page without website that is installed:

  missing module "[root.widget]"

This is caused by the fact that there is no root widget in the frontend
bundle without website installed. This may prevent executing tours
in some cases, because the tour manager needs a root widget to run.

This commit solves this issue by ensuring a root widget in the frontend
even when website is not installed, so that tour manager can run.