[CHG] *: "family" names of admin & demo users
avw/fp? request

3e9f5ccdf38f13033acd3cb097c10249954f66cf changed various demo data,
including putting more "realistic" names in for the administrator and
demo users in demo mode. This is apparently inconvenient when actually
demo-ing Odoo as it's harder to see at a glance whether the current user
is the administrator or "regular" (demo) user of the db, though this
should have less impact since the userification of the admin
(ef79fc854f15759a2c36a1a58aa86a4212d344bc).