[IMP] website: ensure login uniqueness per website in SQL
Constraints in SQL are faster than python equivalent.
Adding a coalesce in PostgreSQL constraint doesn't seems possible.
This combination of constraint and weird named unique index is the
fastest way to ensure a unique constraint with NULL values.

This index require the website_id to be present in the res_users
table. To do this, a stored readonly related field is used.

This was introduced by 1e5a4471378adedd033392f8b177134547267d51