[FIX] hr_timesheet: create project with allow_timesheet checked by default
The default value when creating a project is true for 'allow_timesheet'
field. This will automatically create an Analytic Account (if not
provided). However, the old onchange reset the allow_timesheet to false
if no AA was given, which is a contraction with the expected behavior
on project creation. But, we might want to uncheck 'allow_timesheet'
when we remove the AA on an existing project (Users can not log
timesheets without an active analytic account).
This commit fixes the erasing default value on creation by checking
the record is currently existing in the database.

Task #1886691